bmh line parser added

+++ include/src/nproto/http.h
@@ -24,5 +24,5 @@ enum __em_http_headers {
 
 typedef struct {
 	/* header value, index offset form [0]->[1] */
-	uint8_t headers_range[NP_HTTP_MAX][2];
+	int16_t headers_range[NP_HTTP_MAX][2];
 } nproto_http_t;
